Update deploy.py
Version 2.0 adds support to dynamically rewrite the SDE connection within the toolbox.  I see no way to alter this on the fly other than to find and replace the connection string such that arcpy can then itself find and replace the connection string.  Using anything other than a scalar in the toolbox for the connection seems to limit arcpy from doing its find and replace thing.